adding domain specific language for cursor creationg.
We have a simple parser that works on the head of a list as a cursor description, defaulting to a simle sequence cursor. We also have a packing frameork to take cusors from parsing and make them more logical (right now the simple parser break evey element of a sequence into an individual seq cursor with only one element. The packer for concat cursors packs any such adjaceny sequence cursors into a single one